requirements.txt
asgiref==3.9.1
Django==4.2.23
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
pytest==8.2.1
pytest-django==4.8.0
